A bit less ceremony and add integration test to check it runs without the native jars
