--- src/main/java/io/vertx/core/spi/transport/KQeueTransport.java
@@ -1,109 +1,0 @@
-/*
- * Copyright (c) 2011-2017 The original author or authors
- * ------------------------------------------------------
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * and Apache License v2.0 which accompanies this distribution.
- *
- *     The Eclipse Public License is available at
- *     http://www.eclipse.org/legal/epl-v10.html
- *
- *     The Apache License v2.0 is available at
- *     http://www.opensource.org/licenses/apache2.0.php
- *
- * You may elect to redistribute this code under either of these licenses.
- */
-package io.vertx.core.spi.transport;
-
-import io.netty.channel.Channel;
-import io.netty.channel.ChannelOption;
-import io.netty.channel.EventLoopGroup;
-import io.netty.channel.ServerChannel;
-import io.netty.channel.kqueue.KQueue;
-import io.netty.channel.kqueue.KQueueChannelOption;
-import io.netty.channel.kqueue.KQueueDatagramChannel;
-import io.netty.channel.kqueue.KQueueDomainSocketChannel;
-import io.netty.channel.kqueue.KQueueEventLoopGroup;
-import io.netty.channel.kqueue.KQueueServerDomainSocketChannel;
-import io.netty.channel.kqueue.KQueueServerSocketChannel;
-import io.netty.channel.kqueue.KQueueSocketChannel;
-import io.netty.channel.socket.DatagramChannel;
-import io.netty.channel.socket.InternetProtocolFamily;
-import io.netty.channel.unix.DomainSocketAddress;
-
-import java.net.InetSocketAddress;
-import java.net.SocketAddress;
-import java.util.concurrent.ThreadFactory;
-
-/**
- * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
- */
-class KQeueTransport implements Transport {
-
-  @Override
-  public SocketAddress convert(io.vertx.core.net.SocketAddress address, boolean resolved) {
-    if (address.path() != null) {
-      return new DomainSocketAddress(address.path());
-    } else {
-      if (resolved) {
-        return new InetSocketAddress(address.host(), address.port());
-      } else {
-        return InetSocketAddress.createUnresolved(address.host(), address.port());
-      }
-    }
-  }
-
-  @Override
-  public boolean isAvailable() {
-    return KQueue.isAvailable();
-  }
-
-  @Override
-  public Throwable unavailabilityCause() {
-    return KQueue.unavailabilityCause();
-  }
-
-  @Override
-  public EventLoopGroup eventLoopGroup(int nThreads, ThreadFactory threadFactory, int ioRatio) {
-    KQueueEventLoopGroup eventLoopGroup = new KQueueEventLoopGroup(nThreads, threadFactory);
-    eventLoopGroup.setIoRatio(ioRatio);
-    return eventLoopGroup;
-  }
-
-  @Override
-  public DatagramChannel datagramChannel() {
-    return new KQueueDatagramChannel();
-  }
-
-  @Override
-  public DatagramChannel datagramChannel(InternetProtocolFamily family) {
-    return new KQueueDatagramChannel();
-  }
-
-  @Override
-  public Channel socketChannel(boolean domain) {
-    if (domain) {
-      return new KQueueDomainSocketChannel();
-    } else {
-      return new KQueueSocketChannel();
-    }
-  }
-
-  @Override
-  public ServerChannel serverChannel(boolean domain) {
-    if (domain) {
-      return new KQueueServerDomainSocketChannel();
-    } else {
-      return new KQueueServerSocketChannel();
-    }
-  }
-
-  @Override
-  public ChannelOption<?> channelOption(String name) {
-    switch (name) {
-      case "SO_REUSEPORT":
-        return KQueueChannelOption.SO_REUSEPORT;
-    }
-    return null;
-  }
-}
